Added: EssayQuestion.cs; Exam.cs; ExamMultipleChoiceAnswer.cs; MultipleChoiceCandidateAnswer.cs; Updated: Application.cs; ApplicationDocument.cs; ExamQuestion.cs; ExamSet.cs; MultipleChoiceQuestion.cs; Position.cs; Removed: ApplicantExamAnswer.cs; ApplicationDocument.Category.cs;

+++ LGU/Entities/HumanResource/Position.cs
@@ -1,6 +1,6 @@
 ﻿namespace LGU.Entities.HumanResource
 {
-    public class Position : Entity<long>
+    public class Position : Entity<int>
     {
         public string Description { get; set; }
         public string Abbreviation { get; set; }
